[Refactor] コードの品質及び一貫性改善#335 (#339)
* コードの品質及び一貫性改善

* バッククォートの削除

---------
